Create OverlappingPunctaMeasure
This macro was developed in order to quantify the puncta area versus total cell area. This macro can also use any measure from any of the ROIs made from this thresholding method. This macro is being verified that it can be used for any condition that has two channels that want to be analyzed.